[REFACTOR]하이라이트페이지-수정
